New connected components implementation (#119)
This PR replaces the GraphX implementation of connected components by a DataFrame-based implementation. It also switched the algorithm to the one proposed in "Connected Components in MapReduce and Beyond". The new implementation handles skewness by separating the joins into broadcast join of vertices of high degrees and normal join of the rest.

The implementation was tested on a graph w/ ~2 billion edges, but without long ID conversion.

Note that it uses indexedVertices and indexedEdges to assign long IDs. This might have performance issues if the input graph is skewed. We can do a follow-up PR to address it.

The unit tests are updated to check the broadcast join path. The coverage is apparently not ideal at this time.

Commit list:
* implement the new CC algorithm in DF
* set checkpointDir in test
* return all vertex attributes
* add CC.run back
* remove checkpointInterval
* remove checkpointInterval
* add unit test for broadcast join path
* make it backward compatible with Spark 1.6
* switched back to as[(Long, Long)]
* address comments