[#122] Add option to use GraphX for connected components (#125)
This PR adds a new parameter algorithm to ConnectedComponents, so users can use the implementation in GraphX by setting it to "graphx". The default algorithm is called "graphframes". I thought about giving the algorithm a better name. But the names are either too long, e.g., "alternating-large-star-small-star", or hard to remember, e.g., "kiveris-2014".

Unit tests for friends graph are updated to test both code paths. More unit tests will be added as part of #121.

Commits:
* add option to use GraphX for CC
* make CC test compatible with 1.6.2