[#124] Support checkpoint interval in connected components (#126)
This PR adds checkpointInterval to the new CC implementation. The logic for removing out-of-scope persisted DataFrames was removed to simplify the code. Those should be cleaned by Spark automatically. System.gc() gives hint.

The default value is 1 to avoid exploding the query plan. This might change after we support query plan checkpointing in Spark SQL.

Commits:
* support checkpoint interval
* add unit tests
* fix test in 1.6