Fix negation in motif finding (#278)
This PR fixes the issues discussed in #276

This is the follow-up to #279 which added helper methods to support this main PR.

Previously, we handled each term in the motif in the order given. For each new term, we augmented (a) the set of named elements and (b) the DataFrame representing all instances of the motifs. This caused problems for named elements which only appeared in negated terms.

This implementation does the following:

Separate non-negative terms from negative terms.
Generate a DataFrame X of motif instances for the non-negative terms.
Augment X with named vertices appearing only in negative terms. (This is 1 outer join with the vertex DataFrame per named vertex.)
For each negative term, remove rows from X matching the negative term.
This PR also:

fixes self-loop motifs a->a by adding an extra case into findIncremental()
disallows certain motifs:
"()"
"()-[]->()" and "!()-[]->()"
motifs with reused edge names
adds a bunch of unit tests