[ML-4283]Use flattened columns in motif finding to utilize CBO (#8)
Use flattened columns in motif finding.
In this way, we can utilize the cost-based optimizer to generate a good query plan.

To use cost-based optimizer, you need to set "spark.sql.cbo.enabled" as true, analyze edge and vertex tables in GraphFrames.

If CBO does not work for big patterns. It is usually because the number of joins exceeds the threshold in CBO. You may want to set "spark.sql.cbo.joinReorder.dp.threshold" to a larger number.

Test on the 8-edge-chain query, it decomposes the pattern to two 4-edge-chain patterns and join together. Without CBO, it just joins one edge or vertex each iteration. With CBO, it can generate a good query plan and reduce the time from 15 mins to 30 seconds.