Replace bintray (#385)
Bintray is being deprecated which will cause builds to fail because there are dependencies on Bintray: https://jfrog.com/blog/into-the-sunset-bintray-jcenter-gocenter-and-chartcenter/

Also see #384

This PR replaces the usages of Bintray with repo.scala-sbt.org which I found from this issue on SBT about replacing Bintray. I was able to run a build locally with these changes during the April 26 Bintray brownout.